Fix empty notification switch, restore VR window behaviour; verify to zero damage
Real damage found by tools/verify/dexdiff.py and repaired against the
original bytecode:
- GridConnectionService.notifySettingsByType: jadx emitted an empty
  switch, so group and private IMs used the local-chat LED/sound settings.
- CardboardActivity: the VR runtime refactor dropped keep-screen-on,
  immersive fullscreen, VR mode and sustained-performance mode.
- PhotoViewAttacher.setScale: restore integer halving of the view size.

Nested classes hoisted to top-level files (NavDrawerItem, TransactionViewHolder,
SearchViewHolder, UploadImageResult) changed ButterKnife binding names; their
adapters now ship as original bytecode with the original nesting.

tools/verify/accepted.txt records every remaining difference that was
reviewed and judged intentional (Android 8-14 requirements, render backend
and VR runtime refactors, Room migration, PriorityBinQueue bug fixes) or
d8 codegen-only, each with its reason. verify_against_apk.sh now reports
0 damaged classes and 0 unresolved links; unit tests and protocol
conformance pass.

+++ app/src/main/java/uk/co/senab/photoview/PhotoViewAttacher.java
@@ -856,7 +856,7 @@ public class PhotoViewAttacher implements IPhotoView, View.OnTouchListener, OnGe
         if (imageView == null) {
             return;
         }
-        setScale(f, imageView.getRight() / 2.0f, imageView.getBottom() / 2.0f, z);
+        setScale(f, (float) (imageView.getRight() / 2), (float) (imageView.getBottom() / 2), z);
     }
 
     @Override // uk.co.senab.photoview.IPhotoView
